data_science/cloud_functions/ProcessBattles/requirements.txt
google-cloud-storage==1.26.0
requests==2.23.0